Move a couple of test helpers
