+++ tests/OmniSharp.MSBuild.Tests/WorkspaceInformationTests.cs
@@ -6,8 +6,6 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using OmniSharp.Models;
 using OmniSharp.Models.CodeCheck;
-using OmniSharp.Models.WorkspaceInformation;
-using OmniSharp.MSBuild.Models;
 using TestUtility;
 using Xunit;
 using Xunit.Abstractions;
@@ -27,7 +25,7 @@ namespace OmniSharp.MSBuild.Tests
             using (var testProject = await TestAssets.Instance.GetTestProjectAsync("ProjectAndSolution"))
             using (var host = CreateOmniSharpHost(testProject.Directory))
             {
-                var workspaceInfo = await GetWorkspaceInfoAsync(host);
+                var workspaceInfo = await host.GetMSBuildWorkspaceInfoAsync();
 
                 Assert.Equal("ProjectAndSolution.sln", Path.GetFileName(workspaceInfo.SolutionPath));
                 Assert.NotNull(workspaceInfo.Projects);
@@ -55,7 +53,7 @@ namespace OmniSharp.MSBuild.Tests
             using (var testProject = await TestAssets.Instance.GetTestProjectAsync("ProjectAndSolutionWithProjectSection"))
             using (var host = CreateOmniSharpHost(testProject.Directory))
             {
-                var workspaceInfo = await GetWorkspaceInfoAsync(host);
+                var workspaceInfo = await host.GetMSBuildWorkspaceInfoAsync();
 
                 Assert.Equal("ProjectAndSolutionWithProjectSection.sln", Path.GetFileName(workspaceInfo.SolutionPath));
                 Assert.NotNull(workspaceInfo.Projects);
@@ -71,7 +69,7 @@ namespace OmniSharp.MSBuild.Tests
             using (var testProject = await TestAssets.Instance.GetTestProjectAsync("TwoProjectsWithSolution"))
             using (var host = CreateOmniSharpHost(testProject.Directory))
             {
-                var workspaceInfo = await GetWorkspaceInfoAsync(host);
+                var workspaceInfo = await host.GetMSBuildWorkspaceInfoAsync();
 
                 Assert.Equal("TwoProjectsWithSolution.sln", Path.GetFileName(workspaceInfo.SolutionPath));
                 Assert.NotNull(workspaceInfo.Projects);
@@ -95,7 +93,7 @@ namespace OmniSharp.MSBuild.Tests
             using (var testProject = await TestAssets.Instance.GetTestProjectAsync("ProjectWithGeneratedFile"))
             using (var host = CreateOmniSharpHost(testProject.Directory))
             {
-                var workspaceInfo = await GetWorkspaceInfoAsync(host);
+                var workspaceInfo = await host.GetMSBuildWorkspaceInfoAsync();
 
                 Assert.NotNull(workspaceInfo.Projects);
                 var project = Assert.Single(workspaceInfo.Projects);
@@ -111,7 +109,7 @@ namespace OmniSharp.MSBuild.Tests
             using (var testProject = await TestAssets.Instance.GetTestProjectAsync("ProjectWithSdkProperty"))
             using (var host = CreateOmniSharpHost(testProject.Directory))
             {
-                var workspaceInfo = await GetWorkspaceInfoAsync(host);
+                var workspaceInfo = await host.GetMSBuildWorkspaceInfoAsync();
 
                 Assert.NotNull(workspaceInfo.Projects);
                 var project = Assert.Single(workspaceInfo.Projects);
@@ -127,7 +125,7 @@ namespace OmniSharp.MSBuild.Tests
             using (var testProject = await TestAssets.Instance.GetTestProjectAsync("CSharpAndFSharp"))
             using (var host = CreateOmniSharpHost(testProject.Directory))
             {
-                var workspaceInfo = await GetWorkspaceInfoAsync(host);
+                var workspaceInfo = await host.GetMSBuildWorkspaceInfoAsync();
 
                 Assert.NotNull(workspaceInfo.Projects);
                 var project = Assert.Single(workspaceInfo.Projects);
@@ -143,7 +141,7 @@ namespace OmniSharp.MSBuild.Tests
             using (var testProject = await TestAssets.Instance.GetTestProjectAsync("TwoProjectsWithSolution"))
             using (var host = CreateOmniSharpHost(Path.Combine(testProject.Directory, "App")))
             {
-                var workspaceInfo = await GetWorkspaceInfoAsync(host);
+                var workspaceInfo = await host.GetMSBuildWorkspaceInfoAsync();
 
                 Assert.NotNull(workspaceInfo.Projects);
                 Assert.Equal(2, workspaceInfo.Projects.Count);
@@ -162,7 +160,7 @@ namespace OmniSharp.MSBuild.Tests
             using (var testProject = await TestAssets.Instance.GetTestProjectAsync("SolutionWithSignedProject"))
             using (var host = CreateOmniSharpHost(Path.Combine(testProject.Directory)))
             {
-                var workspaceInfo = await GetWorkspaceInfoAsync(host);
+                var workspaceInfo = await host.GetMSBuildWorkspaceInfoAsync();
                 Assert.NotNull(workspaceInfo.Projects);
                 Assert.Equal(2, workspaceInfo.Projects.Count);
 
@@ -182,7 +180,7 @@ namespace OmniSharp.MSBuild.Tests
                 var token = BitConverter.ToString(AssemblyName.GetAssemblyName(signedProject.TargetPath).GetPublicKeyToken());
                 Assert.Equal("A5-D8-5A-5B-AA-39-A6-A6", token, ignoreCase: true);
 
-                var response = await GetCodeChecksync(host, Path.Combine(testProject.Directory, "CallerLib\\Caller.cs"));
+                var response = await host.CodeCheckRequestAsync(Path.Combine(testProject.Directory, "CallerLib", "Caller.cs"));
                 // Log result to easier debugging of the test should it fail during automated valdiation
                 foreach (var fix in response.QuickFixes)
                 {
@@ -199,7 +197,7 @@ namespace OmniSharp.MSBuild.Tests
             using (var testProject = await TestAssets.Instance.GetTestProjectAsync("ProjectWithMultiTFMLib"))
             using (var host = CreateOmniSharpHost(Path.Combine(testProject.Directory, "App")))
             {
-                var workspaceInfo = await GetWorkspaceInfoAsync(host);
+                var workspaceInfo = await host.GetMSBuildWorkspaceInfoAsync();
 
                 Assert.NotNull(workspaceInfo.Projects);
                 Assert.Equal(2, workspaceInfo.Projects.Count);
@@ -220,7 +218,7 @@ namespace OmniSharp.MSBuild.Tests
             using (var testProject = await TestAssets.Instance.GetTestProjectAsync("AntlrGeneratedFiles"))
             using (var host = CreateOmniSharpHost(testProject.Directory))
             {
-                var workspaceInfo = await GetWorkspaceInfoAsync(host);
+                var workspaceInfo = await host.GetMSBuildWorkspaceInfoAsync();
 
                 Assert.NotNull(workspaceInfo.Projects);
                 var project = Assert.Single(workspaceInfo.Projects);
@@ -228,29 +226,6 @@ namespace OmniSharp.MSBuild.Tests
                 Assert.Equal(6, project.SourceFiles.Count);
                 Assert.Contains(project.SourceFiles, fileName => fileName.EndsWith("GrammarParser.cs"));
             }    
-        }
-
-        private static async Task<MSBuildWorkspaceInfo> GetWorkspaceInfoAsync(OmniSharpTestHost host)
-        {
-            var service = host.GetWorkspaceInformationService();
-
-            var request = new WorkspaceInformationRequest
-            {
-                ExcludeSourceFiles = false
-            };
-
-            var response = await service.Handle(request);
-
-            return (MSBuildWorkspaceInfo)response["MsBuild"];
-        }
-
-        private static async Task<QuickFixResponse> GetCodeChecksync(OmniSharpTestHost host, string filePath)
-        {
-            var service = host.GetCodeCheckServiceService();
-
-            var request = new CodeCheckRequest { FileName = filePath };
-
-            return await service.Handle(request);
         }
     }
 }
